+++ lrpar/src/lib/mod.rs
@@ -26,7 +26,7 @@
 //!
 //! ```text
 //! use cfgrammar::yacc::YaccKind;
-//! use lrlex::LexerBuilder;
+//! use lrlex::CTLexerBuilder;
 //! use lrpar::CTParserBuilder;
 //!
 //! fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -34,7 +34,7 @@
 //!         .yacckind(YaccKind::Grmtools)
 //!         .grammar_in_src_dir("calc.y")?
 //!         .build()?;
-//!     LexerBuilder::new()
+//!     CTLexerBuilder::new()
 //!         .rule_ids_map(cp.lexeme_id_map())
 //!         .lexer_in_src_dir("calc.l")?
 //!         .build()?;